Update max31865.c
The threshold register value was not been written correctly. The most significant bits come first at buff[0] with a proper bit rotation to the right then the lest significant bits.
In the reading ADC register value, all of the 16-bits value was been rotated to the right instead of the low significat half(buff[1]).. With this changes the trheshold callback is triggered when there is not sensor attached to the chip. And the readings are correct